[packages/template,core][m]: Icons, Nav and Search moved
