+++ package.json
@@ -75,6 +75,7 @@
   },
   "dependencies": {
     "core-js": "^3.6.5",
+    "next": "^13.0.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "regenerator-runtime": "0.13.7",
