+++ packages/core/src/ui/Base/BaseLink.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import { forwardRef } from "react";
 
-const BaseLink = forwardRef((props, ref) => {
+const BaseLink = forwardRef((props: any, ref) => {
   const { href, children, ...rest } = props;
   return (
     <Link href={href} ref={ref} {...rest}>
