+++ packages/core/src/ui/Base/index.ts
@@ -1,0 +1,2 @@
+export { BaseLink } from "./BaseLink";
+export { ThemeSelector } from "./ThemeSelector";
