+++ packages/core/src/ui/Icons/index.tsx
@@ -1,0 +1,5 @@
+export { GitHubIcon } from "./GitHubIcon";
+export { DiscordIcon } from "./DiscordIcon";
+export { MenuIcon } from "./MenuIcon";
+export { CloseIcon } from "./CloseIcon";
+export { SearchIcon } from "./SearchIcon";
