--- packages/core/src/ui/Nav/Nav.module.scss
@@ -1,7 +1,0 @@
-/*
- * Replace this with your own classes
- *
- * e.g.
- * .container {
- * }
-*/
