--- packages/core/src/ui/Nav/Nav.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from "@testing-library/react";
-
-import Nav from "./Nav";
-
-describe("Nav", () => {
-  it("should render successfully", () => {
-    const { baseElement } = render(<Nav />);
-    expect(baseElement).toBeTruthy();
-  });
-});
