+++ packages/core/src/ui/Search/AlgoliaModal.jsx
