+++ packages/core/src/ui/Search/index.ts
@@ -1,0 +1,3 @@
+// TODO tidy up this API
+export { SearchField } from "./SearchField";
+export { SearchContext } from "./SearchProvider";
