+++ packages/core/src/ui/index.ts
@@ -1,1 +1,2 @@
 export { Nav } from "./Nav";
+export * from "./Icons";
