+++ packages/template/package.json
@@ -10,9 +10,7 @@
     "export": "npm run build && next export"
   },
   "dependencies": {
-    "@docsearch/react": "^3.2.0",
     "@flowershow/core": "workspace:*",
-    "@headlessui/react": "^1.6.6",
     "@heroicons/react": "^1.0.4",
     "@silvenon/remark-smartypants": "^1.0.0",
     "@tailwindcss/typography": "^0.5.2",
@@ -21,13 +19,11 @@
     "date-fns": "^2.29.2",
     "gray-matter": "^4.0.3",
     "hastscript": "^7.0.2",
-    "kbar": "^0.1.0-beta.37",
     "mdx-mermaid": "^2.0.0-rc1",
     "mermaid": "^9.1.4",
     "next": "^13.0.4",
     "next-contentlayer": "^0.2.9",
     "next-seo": "^5.14.1",
-    "next-themes": "^0.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "rehype-autolink-headings": "^6.1.1",
